extract text translation methods to new file
